pyproject.toml
[project]
name = "neo-assetmaker"
version = "2.0.0"
description = "明日方舟通行证素材制作器"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PyQt6>=6.5.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "jsonschema>=4.17.0",
    "thefuzz>=0.22.0",
    "python-Levenshtein>=0.25.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[dependency-groups]
dev = [
    "cx_Freeze>=6.15.0",
]
